chore: extract
